refactor: Move `EnemySpawnMgrOld` to separate file
